% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/queryogcio.R
\name{queryogcio}
\alias{queryogcio}
\title{Query OGCIO API to parse free text address}
\usage{
queryogcio(text_address, n = 1, sleep_time = 0)
}
\arguments{
\item{text_address}{String of Hong Kong Address}

\item{n}{Integer, Number of suggested address to return}

\item{sleep_time}{Numeric, sleep time between each query}
}
\value{
A data.frame of parsed address(es)
}
\description{
This function query the OGCIO server and return a parsed address.
}
\examples{
queryogcio("66 Causeway Road, Causeway Bay, Hong Kong")
}

